remove(pack): Fruits delight
This mod is cluttersome and feels incomplete. The fruits grow in the leaves of vanilla trees, which is confusing, and they yield too much food too fast. I kept it for the jello mixing in vanilla cauldrons but I've never gotten any use out of it and I'd rather streamline the pack by removing the clutter and reducing food options than further evaluate this.